chore(refactor): refactor webpack/config/bundles.js
