+++ packages/holocron-dev-server/src/webpack/configs/fragments.js
@@ -13,115 +13,10 @@
  */
 
 import path from 'path';
-import { BundleAnalyzerPlugin } from 'webpack-bundle-analyzer';
-import {
-  HotModuleReplacementPlugin,
-  DllReferencePlugin,
-  EnvironmentPlugin,
-  DefinePlugin,
-} from 'webpack';
 
-import { modulesBundleName, externalsBundleName } from '../../constants';
 import {
   getContextPath,
-  getModulesPath,
-  getVendorsPath,
-  getPublicModulesUrl,
-  createModuleScriptUrl,
-  getReportFilename,
 } from '../../utils/paths';
-import HolocronModulePlugin from '../plugins/holocron-webpack-plugin';
-import {
-  modulesLibraryVarName,
-  externalsLibraryVarName,
-  getWebpackVersion,
-  createOneAppExternals,
-  createHolocronModuleEntries,
-} from '../helpers';
-import { isDevelopment } from '../../utils/helpers';
-
-export function createDllReferenceConfigFragment({ name = externalsBundleName } = {}) {
-  return {
-    plugins: [
-      new DllReferencePlugin({
-        context: getContextPath(),
-        name: externalsLibraryVarName,
-        manifest: getVendorsPath(`${name}.dll.json`),
-      }),
-    ],
-  };
-}
-
-export function createHolocronModulesConfigFragment({ modules, externals, hot } = {}) {
-  const webpackVersion = getWebpackVersion();
-  const libKey = webpackVersion >= 5 ? 'uniqueName' : 'library';
-  const hashName = webpackVersion >= 5 ? 'fullhash' : 'hash';
-  const fragment = {
-    entry: createHolocronModuleEntries({ modules, hot }),
-    externals: createOneAppExternals(),
-    output: {
-      publicPath: getPublicModulesUrl(),
-      path: getModulesPath(),
-      filename: createModuleScriptUrl('[name]'),
-      [libKey]: modulesLibraryVarName,
-    },
-    optimization: {
-      runtimeChunk: 'single',
-    },
-    plugins: [
-      new HolocronModulePlugin({
-        modules,
-        externals,
-        hot,
-      }),
-    ],
-  };
-
-  // base hot module reload with webpack
-  if (hot) {
-    fragment.plugins.push(new HotModuleReplacementPlugin());
-    Object.assign(fragment.output, {
-      hotUpdateChunkFilename: `[name]/[id].[${hashName}].hot-update.js`,
-      hotUpdateMainFilename: `__hot/[${hashName}].hot-update.json`,
-    });
-  }
-
-  return fragment;
-}
-
-export function createBundleAnalyzerConfigFragment({ name = modulesBundleName } = {}) {
-  return {
-    plugins: [
-      new BundleAnalyzerPlugin({
-        openAnalyzer: false,
-        generateStatsFile: false,
-        logLevel: 'silent',
-        analyzerMode: 'static',
-        reportFilename: `./${getReportFilename(name)}`,
-      }),
-    ],
-  };
-}
-
-export function createEnvironmentDefinitionsConfigFragment({
-  isDev = isDevelopment(),
-  environmentVariables = {},
-  globalDefinitions = {},
-  browser = true,
-} = {}) {
-  return {
-    plugins: [
-      new EnvironmentPlugin({
-        ...environmentVariables,
-        NODE_ENV: isDev ? 'development' : 'production',
-      }),
-      new DefinePlugin({
-        ...globalDefinitions,
-        'global.BROWSER': JSON.stringify(browser),
-      }),
-    ],
-  };
-}
 
 export function createResolverConfigFragment({
   context = getContextPath(),
